Criado modo dev e prod e inciado crud do Address, createAddress

+++ src/main/java/com/federal/prision/domain/Person.java
@@ -5,7 +5,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 import jakarta.persistence.CascadeType;
@@ -20,7 +19,7 @@ public class Person {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String socialSecurity;
 	
 	private LocalDate birthDate;
@@ -35,7 +34,7 @@ public class Person {
 	public Person() {
 	}
 
-	public Person(Integer id, String socialSecurity, LocalDate birthDate, String name, String email) {
+	public Person(Long id, String socialSecurity, LocalDate birthDate, String name, String email) {
 		this.id = id;
 		this.socialSecurity = socialSecurity;
 		this.birthDate = birthDate;
@@ -43,11 +42,11 @@ public class Person {
 		this.email = email;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
